requirements.txt
langchain==0.3.17
langchain-core==0.3.59
langchain-community==0.3.16
langchain-ollama==0.3.2
langchain-google-genai==2.1.3
langchain-groq==0.3.2
chromadb==1.0.10
langchain-huggingface==0.2.0
langchain_text_splitters==0.3.5
langsmith==0.3.4
langgraph==0.3.31
sentence-transformers==4.1.0
protobuf==5.29.4
fastapi==0.115.9
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic-settings==2.9.1
pydantic==2.11.3
openai-whisper
pymupdf==1.26.0
gradio==5.32.0
httpx==0.28.1